move builtins from pyleus.compat to pyleus.testing
builtins is only used in mocking builtins.open in the testing code
