+++ tests/cli/virtualenv_proxy_test.py
@@ -6,8 +6,7 @@ import pytest
 from pyleus import exception
 from pyleus.cli import virtualenv_proxy
 from pyleus.cli.virtualenv_proxy import VirtualenvProxy
-from pyleus.compat import builtins
-from pyleus.testing import mock
+from pyleus.testing import mock, builtins
 
 
 VENV_PATH = "/tmp/my/beloved/venv"
